+++ gw_spaceheat/actors/primary_scada/primary_scada.py
@@ -1,7 +1,7 @@
 from typing import List, Dict
 from actors.primary_scada.primary_scada_base import PrimaryScadaBase
 from data_classes.sh_node import ShNode
-from data_classes.boolean_actuator_component import BooleanActuatorComponent 
+from data_classes.components.boolean_actuator_component import BooleanActuatorComponent 
 from drivers.boolean_actuator.boolean_actuator_base import BooleanActuator
 from schema.gt.gt_telemetry.gt_telemetry_1_0_0_maker import GtTelemetry100, GtTelemetry100_Maker
 from schema.gs.gs_pwr_1_0_0_maker import GsPwr100_Maker, GsPwr100
