+++ gw_spaceheat/data_classes/components/electric_heater_component.py
@@ -1,22 +1,22 @@
 from typing import Optional
 
+from data_classes.cacs.electric_heater_cac import ElectricHeaterCac
 from data_classes.component import Component
-from data_classes.boolean_actuator_cac import BooleanActuatorCac
-from data_classes.errors import DcError, DataClassLoadingError
+from data_classes.errors import DataClassLoadingError, DcError
 
-class BooleanActuatorComponent(Component):
+
+class ElectricHeaterComponent(Component):
     by_id = {}
     
     base_props = []
     base_props.append('component_id')
     base_props.append('display_name')
     base_props.append('component_attribute_class_id')
-    base_props.append('gpio')
 
     def __new__(cls, component_id, *args, **kwargs):
         if component_id in Component.by_id.keys():
             if not isinstance(Component.by_id[component_id], cls):
-                raise Exception(f"Id already exists, not an actuator!")
+                raise Exception(f"Id already exists, not a sensor!")
             return Component.by_id[component_id]
         instance = super().__new__(cls,component_id=component_id)
         Component.by_id[component_id] = instance
@@ -25,15 +25,13 @@ class BooleanActuatorComponent(Component):
     def __init__(self,
                  component_id: Optional[str] = None,
                  display_name: Optional[str] = None,
-                 component_attribute_class_id: Optional[str] = None,
-                 gpio: Optional[int] = None):
-        super(BooleanActuatorComponent, self).__init__(component_id=component_id,
+                 component_attribute_class_id: Optional[str] = None):
+        super(ElectricHeaterComponent, self).__init__(component_id=component_id,
                             display_name=display_name,
                             component_attribute_class_id=component_attribute_class_id)
-        self.gpio= gpio
 
     def __repr__(self):
-        return f'Component {self.display_name} => Cac {self.cac.display_name}. GPIO: {self.gpio}'
+        return f'Component {self.display_name} => Cac {self.cac.display_name}'
 
     @classmethod
     def check_uniqueness_of_primary_key(cls, attributes):
@@ -51,17 +49,11 @@ class BooleanActuatorComponent(Component):
 
     @classmethod
     def check_initialization_consistency(cls, attributes):
-        BooleanActuatorComponent.check_uniqueness_of_primary_key(attributes)
-        BooleanActuatorComponent.check_existence_of_certain_attributes(attributes)
+        ElectricHeaterComponent.check_uniqueness_of_primary_key(attributes)
+        ElectricHeaterComponent.check_existence_of_certain_attributes(attributes)
 
     @property
-    def cac(self) -> BooleanActuatorCac:
-        if self.component_attribute_class_id not in BooleanActuatorCac.by_id.keys():
-            raise DataClassLoadingError(f"BooleanActuatorCacId {self.component_attribute_class_id} not loaded yet")
-        return BooleanActuatorCac.by_id[self.component_attribute_class_id]
-
-    @property
-    def make_model(self) -> str:
-        if self.component_attribute_class_id not in BooleanActuatorCac.by_id.keys():
-            raise DataClassLoadingError(f"BooleanActuatorCacId {self.component_attribute_class_id} not loaded yet")
-        return f'{BooleanActuatorCac.by_id[self.component_attribute_class_id].make_model}'
+    def cac(self) -> ElectricHeaterCac:
+        if self.component_attribute_class_id not in ElectricHeaterCac.by_id.keys():
+            raise DataClassLoadingError(f"ElectricHeaterCacId {self.component_attribute_class_id} not loaded yet")
+        return ElectricHeaterCac.by_id[self.component_attribute_class_id]
