+++ gw_spaceheat/data_classes/sensor_type.py
@@ -1,7 +1,8 @@
 """ SensorType Class Definition """
-from typing import Optional
 from abc import ABC
-from .mixin import StreamlinedSerializerMixin
+from typing import Optional
+
+from data_classes.mixin import StreamlinedSerializerMixin
 
 
 class SensorType(ABC, StreamlinedSerializerMixin):
