+++ gw_spaceheat/data_classes/sh_node_base.py
@@ -1,8 +1,8 @@
 """ ShNode Base Class Definition """
 
-from typing import Optional
 from abc import ABC, abstractproperty
-from typing import List
+from typing import List, Optional
+
 from data_classes.mixin import StreamlinedSerializerMixin
