+++ gw_spaceheat/data_classes/sh_node_role.py
@@ -1,8 +1,9 @@
 """ ShNodeRole Definition """
 
-from typing import Optional
 from abc import ABC
-from .mixin import StreamlinedSerializerMixin
+from typing import Optional
+
+from data_classes.mixin import StreamlinedSerializerMixin
 
 
 class ShNodeRole(ABC, StreamlinedSerializerMixin):
