+++ gw_spaceheat/data_classes/thermal_edge.py
@@ -1,5 +1,7 @@
 from typing import NamedTuple
+
 from data_classes.sh_node import ShNode
+
 
 class ThermalEdge(NamedTuple):
     FromNode: ShNode
