+++ gw_spaceheat/data_classes/time_step_base.py
@@ -1,8 +1,9 @@
 """ TimeStep Base Class Definition """
 import time
 import uuid
-from typing import Optional
 from abc import ABC, abstractproperty
+from typing import Optional
+
 from data_classes.mixin import StreamlinedSerializerMixin
