+++ gw_spaceheat/data_classes/world_instance_base.py
@@ -1,8 +1,9 @@
 """ WorldInstance Base Class Definition """
 import time
 import uuid
-from typing import Optional
 from abc import ABC, abstractproperty
+from typing import Optional
+
 from data_classes.mixin import StreamlinedSerializerMixin
