+++ gw_spaceheat/requirements/dev.in
@@ -3,3 +3,4 @@
 -r drivers.in
 ipython==8.0.0
 pip-tools
+isort
