+++ gw_spaceheat/requirements/dev.txt
@@ -22,6 +22,8 @@ executing==0.8.3
     # via stack-data
 ipython==8.0.0
     # via -r requirements/dev.in
+isort==5.10.1
+    # via -r requirements/dev.in
 jedi==0.18.1
     # via ipython
 matplotlib-inline==0.1.3
@@ -60,6 +62,8 @@ pygments==2.12.0
     # via ipython
 python-dateutil==2.8.2
     # via pendulum
+python-dotenv==0.20.0
+    # via -r requirements/base.in
 pytz==2022.1
     # via -r requirements/base.in
 pytzdata==2020.1
